Use godeps
I really wanted to use git subtree/_vendor but it wasn't really working
out well this seems like a good comprimise.  If the changes in goirc in
the 'master' branch lands in 'go1' branch may just get rid of
dependency management all together and go back to 'go get'.

<rant type="git commit message">
Go dependency management kinda is in hard area as the community is so
split with way way way to many tools. I'd like it if there was an
official framework from Go to unify the community and build tools on
top of that
</rant>